src/Elksoft.Converters/Elksoft.Converters.StringConverters/TimeSpanToStringConverter.cs
// <copyright file="TimeSpanToStringConverter.cs" company="Georgy Losenkov">
// Copyright (c) Georgy Losenkov. All rights reserved.
// </copyright>

using System;

namespace Elksoft.Converters.StringConverters
{
    internal sealed class TimeSpanToStringConverter : Converter<TimeSpan, String>
    {
        public override Boolean AcceptsNull => false;

        public override Boolean IsExplicit => true;

        public override String? Convert(TimeSpan value, IFormatProvider? formatProvider)
        {
            return value.ToString(null, formatProvider);
        }
    }
}
